requirements.txt.txt
streamlit
rasterio
pillow
folium
streamlit-folium
